requirements.txt
streamlit
anthropic
psycopg2-binary
pytest
streamlit-feedback
